+++ server/src/main/java/org/opensearch/index/compositeindex/datacube/startree/index/StarTreeValues.java
@@ -9,9 +9,12 @@
 package org.opensearch.index.compositeindex.datacube.startree.index;
 
 import org.apache.lucene.codecs.DocValuesProducer;
+import org.apache.lucene.index.DocValues;
+import org.apache.lucene.index.DocValuesType;
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.SegmentReadState;
 import org.apache.lucene.index.SortedNumericDocValues;
+import org.apache.lucene.index.SortedSetDocValues;
 import org.apache.lucene.store.IndexInput;
 import org.opensearch.common.annotation.ExperimentalApi;
 import org.opensearch.index.compositeindex.CompositeIndexMetadata;
@@ -25,6 +28,7 @@ import org.opensearch.index.compositeindex.datacube.startree.fileformats.meta.St
 import org.opensearch.index.compositeindex.datacube.startree.node.StarTreeFactory;
 import org.opensearch.index.compositeindex.datacube.startree.node.StarTreeNode;
 import org.opensearch.index.compositeindex.datacube.startree.utils.iterator.SortedNumericStarTreeValuesIterator;
+import org.opensearch.index.compositeindex.datacube.startree.utils.iterator.SortedSetStarTreeValuesIterator;
 import org.opensearch.index.compositeindex.datacube.startree.utils.iterator.StarTreeValuesIterator;
 
 import java.io.IOException;
@@ -35,7 +39,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Supplier;
 
-import static org.opensearch.index.codec.composite.composite912.Composite912DocValuesReader.getSortedNumericDocValues;
 import static org.opensearch.index.compositeindex.CompositeIndexConstants.SEGMENT_DOCS_COUNT;
 import static org.opensearch.index.compositeindex.CompositeIndexConstants.STAR_TREE_DOCS_COUNT;
 import static org.opensearch.index.compositeindex.datacube.startree.utils.StarTreeUtils.fullyQualifiedFieldNameForStarTreeDimensionsDocValues;
@@ -128,8 +131,15 @@ public class StarTreeValues implements CompositeIndexValues {
 
         // build dimensions
         List<Dimension> readDimensions = new ArrayList<>();
-        for (String dimension : starTreeMetadata.getDimensionFields()) {
-            readDimensions.add(new ReadDimension(dimension));
+        for (String dimension : starTreeMetadata.getDimensionFields().keySet()) {
+            readDimensions.add(
+                new ReadDimension(
+                    dimension,
+                    readState.fieldInfos.fieldInfo(
+                        fullyQualifiedFieldNameForStarTreeDimensionsDocValues(starTreeMetadata.getCompositeFieldName(), dimension)
+                    ).getDocValuesType()
+                )
+            );
         }
 
         // star-tree field
@@ -151,19 +161,25 @@ public class StarTreeValues implements CompositeIndexValues {
         metricValuesIteratorMap = new LinkedHashMap<>();
 
         // get doc id set iterators for dimensions
-        for (String dimension : starTreeMetadata.getDimensionFields()) {
+        for (String dimension : starTreeMetadata.getDimensionFields().keySet()) {
             dimensionValuesIteratorMap.put(dimension, () -> {
                 try {
-                    SortedNumericDocValues dimensionSortedNumericDocValues = null;
+                    FieldInfo dimensionfieldInfo = null;
                     if (readState != null) {
-                        FieldInfo dimensionfieldInfo = readState.fieldInfos.fieldInfo(
+                        dimensionfieldInfo = readState.fieldInfos.fieldInfo(
                             fullyQualifiedFieldNameForStarTreeDimensionsDocValues(starTreeField.getName(), dimension)
                         );
-                        if (dimensionfieldInfo != null) {
-                            dimensionSortedNumericDocValues = compositeDocValuesProducer.getSortedNumeric(dimensionfieldInfo);
-                        }
                     }
-                    return new SortedNumericStarTreeValuesIterator(getSortedNumericDocValues(dimensionSortedNumericDocValues));
+                    assert dimensionfieldInfo != null;
+                    if (dimensionfieldInfo.getDocValuesType().equals(DocValuesType.SORTED_SET)) {
+                        SortedSetDocValues dimensionSortedSetDocValues = compositeDocValuesProducer.getSortedSet(dimensionfieldInfo);
+                        return new SortedSetStarTreeValuesIterator(getSortedSetDocValues(dimensionSortedSetDocValues));
+                    } else {
+                        SortedNumericDocValues dimensionSortedNumericDocValues = compositeDocValuesProducer.getSortedNumeric(
+                            dimensionfieldInfo
+                        );
+                        return new SortedNumericStarTreeValuesIterator(getSortedNumericDocValues(dimensionSortedNumericDocValues));
+                    }
                 } catch (IOException e) {
                     throw new RuntimeException("Error loading dimension StarTreeValuesIterator", e);
                 }
@@ -271,5 +287,31 @@ public class StarTreeValues implements CompositeIndexValues {
 
     public int getStarTreeDocumentCount() {
         return starTreeMetadata.getStarTreeDocCount();
+    }
+
+    /**
+     * Returns the sorted numeric doc values for the given sorted numeric field.
+     * If the sorted numeric field is null, it returns an empty doc id set iterator.
+     * <p>
+     * Sorted numeric field can be null for cases where the segment doesn't hold a particular value.
+     *
+     * @param sortedNumeric the sorted numeric doc values for a field
+     * @return empty sorted numeric values if the field is not present, else sortedNumeric
+     */
+    static SortedNumericDocValues getSortedNumericDocValues(SortedNumericDocValues sortedNumeric) {
+        return sortedNumeric == null ? DocValues.emptySortedNumeric() : sortedNumeric;
+    }
+
+    /**
+     * Returns the sortedSet doc values for the given sortedSet field.
+     * If the sortedSet field is null, it returns an empty doc id set iterator.
+     * <p>
+     * SortedSet field can be null for cases where the segment doesn't hold a particular value.
+     *
+     * @param sortedSetDv the sortedSet doc values for a field
+     * @return empty sortedSet values if the field is not present, else sortedSetDv
+     */
+    static SortedSetDocValues getSortedSetDocValues(SortedSetDocValues sortedSetDv) {
+        return sortedSetDv == null ? DocValues.emptySortedSet() : sortedSetDv;
     }
 }